# docs/swagger.yaml
basePath: /
definitions:
  models.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  models.GetPaymentResponse:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      payment_status:
        type: string
    type: object
  models.PostPaymentRequest:
    properties:
      amount:
        type: integer
      card_number:
        type: string
      currency:
        type: string
      cvv:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
    type: object
  models.PostPaymentResponse:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      payment_status:
        type: string
    type: object
  models.RejectedResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      payment_status:
        type: string
    type: object
host: localhost:8090
info:
  contact: {}
  description: Interview challenge for building a Payment Gateway - Go version
  title: Payment Gateway Challenge Go
paths:
  /api/payments:
    post:
      consumes:
      - application/json
      description: Validates a card payment and forwards it to the acquiring bank
      parameters:
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/models.PostPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PostPaymentResponse'
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/models.RejectedResponse'
        "502":
          description: bank unavailable
          schema:
            type: string
      summary: Process a payment
      tags:
      - payments
  /api/payments/{id}:
    get:
      description: Retrieves a previously processed payment by its id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetPaymentResponse'
        "404":
          description: payment not found
          schema:
            type: string
      summary: Get a payment
      tags:
      - payments
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"